requirements.txt
aiohttp==3.7.3
antiorm==1.2.1
apiclient==1.0.4
argon2-cffi==20.1.0
arrow==1.0.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
cycler==0.10.0
Cython==0.29.17
dataframe-image==0.1.1
db==0.1.1
db-sqlite3==0.0.1
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
et-xmlfile==1.0.1
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-cloud-bigquery==2.13.1
google-cloud-core==1.6.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.37.0
httplib2==0.18.1
idna==2.10
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jira==2.0.0
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
kaleido==0.0.3.post1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
multidict==5.0.2
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numexpr==2.7.1
numpy==1.19.4
O365==2.0.13
oauth2client==4.1.3
oauthlib==3.1.0
openpyxl==3.0.5
orca==1.5.4
packaging==20.7
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
patsy==0.5.1
pbr==5.5.1
pdfkit==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
plotly==4.13.0
pmdarima==1.8.0
prometheus-client==0.9.0
prompt-toolkit==3.0.8
proto-plus==1.18.1
protobuf==3.14.0
psutil==5.7.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.2
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
requests==2.25.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
retrying==1.3.3
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.4
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
statsmodels==0.12.1
stringcase==1.2.0
tables==3.6.1
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==1.2.0
yarl==1.6.3
